Look for arrivals/departures out to two hours, not just one
This allows us to find trips that depart soon but arrive after
the one hour window. Still isn't really correct, and we're now
finding too many trips instead of too few, but it's better than
it was.

A more accurate approach might be to use the old 60 minute
limit but determine trips by fetching trip-details for each
trip from a source stop and checking for destination stops,
but that would likely be quite a bit more expensive.